01-gradio-demo/requirements.txt
pip==24.0
friendli-client==1.3.4
gradio==4.26.0
